requirements/common.txt
django-chart-tools==0.2.1
django-endless-pagination==2.0
django-model-utils==2.2
django-mptt==0.6.1
django-photologue==3.1.1
django-pjaxr==1.0.1
django-widget-tweaks==1.3
docutils==0.7
icalendar==3.8.4
sorl-thumbnail==11.12.1b
simplejson==3.6.5
django-summernote==0.6.0
